Fix a few issues and add some logging in case something breaks
Added a check to see if all of the users have a bank account, because if
they don't then they can't receive credits. Also added a check to see if
the user is trying to split less than 2 credits, because that is
impossible since the bank doesn't deal in fractions of credits. Added
console logging about the split in case something breaks, or someone
does something that they're not supossed to. That way it'll be easier to
revert changes to users' banks.