AnnoyOTron!! Now you get to have fun annoying people on Discord.
AnnoyOTron will auto ping any number of users every minute, for a
maximum of 30 minutes. Very useful if you have forgetful members in your
Discord server that need reminders to do stuff. Has a removeuser
function that will remove a specified user. Users will also be removed
from the pinglist if they send a message in the channel that AOT is
active in. Lastly it has a stop function for if you need to stop it
completely.