New cog - huntingcheck.
huntingcheck keeps an eye on the hunting cog. The hunting cog can act up
if there is any kind of connection issue on Discord's servers or the
bot's server. If it notices that the next hunt is a long time away it
will reload the hunting cog to reset it.